Add softmax into the FCN model

+++ src/utils.py
@@ -15,5 +15,5 @@ def cross_entropy_loss(x, y):
 # Accuracy metric.
 def accuracy(y_pred, y_true):
     # Predicted class is the index of highest score in prediction vector (i.e. argmax).
-    correct_prediction = tf.equal(tf.argmax(y_pred, 1), tf.cast(y_true, tf.int64))
+    correct_prediction = tf.equal(tf.argmax(y_pred, 1), y_true)
     return tf.reduce_mean(tf.cast(correct_prediction, tf.float32), axis=-1)
